refactor(sandbox): move the shared stylesheet into shared/src
