+++ sandbox/react/src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client'
 import { App } from './app'
 
 // The stylesheet both web sandboxes share.
-import '../../shared/styles.css'
+import '../../shared/src/styles.css'
 
 const root = document.getElementById('root')
 if (!root) throw new Error('missing #root')
