+++ sandbox/solid/src/main.tsx
@@ -2,7 +2,7 @@ import { render } from '@solidjs/web'
 import { App } from './app'
 
 // The stylesheet both web sandboxes share.
-import '../../shared/styles.css'
+import '../../shared/src/styles.css'
 
 const root = document.getElementById('root')
 if (!root) throw new Error('missing #root')
